# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 1.5.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /initiatives:
    get:
      operationId: listInitiatives
      tags: [initiatives]
      summary: List all initiatives
      responses:
        "200":
          description: A list of initiatives
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Initiative"
    post:
      operationId: createInitiative
      tags: [initiatives]
      summary: Create a new initiative
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InitiativeInput"
      responses:
        "201":
          description: Created initiative
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Initiative"
  /initiatives/{id}:
    get:
      operationId: getInitiative
      tags: [initiatives]
      summary: Get a single initiative
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: The initiative
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Initiative"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    patch:
      operationId: updateInitiative
      tags: [initiatives]
      summary: Update an initiative
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InitiativeUpdate"
      responses:
        "200":
          description: Updated initiative
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Initiative"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    delete:
      operationId: deleteInitiative
      tags: [initiatives]
      summary: Delete an initiative
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /initiatives/{id}/versions:
    get:
      operationId: listInitiativeVersions
      tags: [initiatives]
      summary: List version history for an initiative
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Version history entries, newest first
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/InitiativeVersion"
  /initiatives/{id}/recommendations:
    get:
      operationId: getInitiativeRecommendations
      tags: [initiatives]
      summary: Get Initiative Intelligence recommendations
      description: >
        Deterministic, rule-based recommendations generated by the
        Initiative Intelligence Engine. The engine sits behind a provider
        abstraction so it can later be replaced by an AI-backed provider
        without changing this contract.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Generated recommendations for the initiative
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InitiativeRecommendations"
        "400":
          description: Invalid id
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /initiatives/{id}/recalculate:
    post:
      operationId: recalculateInitiative
      tags: [initiatives]
      summary: Recalculate scoring, AI readiness, and priority from stored fields
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Recalculation result with a detailed per-component change explanation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RecalculationResult"
        "400":
          description: Invalid id
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /initiatives/{id}/calculations:
    get:
      operationId: listCalculationEvents
      tags: [initiatives]
      summary: List the recalculation audit history for an initiative
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Calculation events, newest first
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CalculationEvent"
        "400":
          description: Invalid id
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /initiatives/{id}/compare:
    get:
      operationId: compareInitiativeVersions
      tags: [initiatives]
      summary: Compare the current version with the previous version
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Side-by-side field comparison of current vs previous version
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VersionComparison"
        "400":
          description: Invalid id
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
components:
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    Initiative:
      type: object
      required:
        - id
        - title
        - jiraLinks
        - department
        - submitterName
        - category
        - status
        - executiveSummary
        - problemStatement
        - currentProcess
        - desiredOutcome
        - aiConcept
        - prototypeGoal
        - successMetric
        - estimatedHoursSavedMonthly
        - estimatedRevenueOpportunity
        - estimatedCostSavings
        - customerImpact
        - complianceRisk
        - technicalComplexity
        - aiReadiness
        - businessValue
        - revenuePotential
        - costSavingsScore
        - customerImpactScore
        - strategicAlignment
        - aiReadinessScore
        - prototypeConfidence
        - technicalComplexityPenalty
        - riskPenalty
        - score
        - priority
        - version
        - assignedTeam
        - currentPhase
        - prototypeDay
        - lastReviewedAt
        - nextReviewAt
        - createdAt
        - updatedAt
      properties:
        id:
          type: integer
        title:
          type: string
        jiraLinks:
          type: array
          items: { $ref: "#/components/schemas/InitiativeJiraLink" }
        department:
          type: string
        submitterName:
          type: string
        businessOwner:
          type: ["string", "null"]
        executiveSponsor:
          type: ["string", "null"]
        category:
          type: string
        status:
          type: string
        executiveSummary:
          type: ["string", "null"]
        problemStatement:
          type: string
        currentProcess:
          type: string
        desiredOutcome:
          type: string
        aiConcept:
          type: string
        prototypeGoal:
          type: string
        successMetric:
          type: string
        estimatedHoursSavedMonthly:
          type: number
        estimatedRevenueOpportunity:
          type: number
        estimatedCostSavings:
          type: number
        customerImpact:
          type: string
        complianceRisk:
          type: string
        technicalComplexity:
          type: string
        aiReadiness:
          type: string
        businessValue:
          type: integer
        revenuePotential:
          type: integer
        costSavingsScore:
          type: integer
        customerImpactScore:
          type: integer
        strategicAlignment:
          type: integer
        aiReadinessScore:
          type: integer
        prototypeConfidence:
          type: integer
        technicalComplexityPenalty:
          type: integer
        riskPenalty:
          type: integer
        score:
          type: integer
        priority:
          type: string
        version:
          type: string
        assignedTeam:
          type: ["string", "null"]
        currentPhase:
          type: ["string", "null"]
        prototypeDay:
          type: ["integer", "null"]
        lastReviewedAt:
          type: ["string", "null"]
        nextReviewAt:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
...
    InitiativeVersion:
      type: object
      required:
        - id
        - initiativeId
        - version
        - changedBy
        - summary
        - createdAt
      properties:
        id:
          type: integer
        initiativeId:
          type: integer
        version:
          type: string
        changedBy:
          type: string
        summary:
          type: string
        createdAt:
          type: string
...
    CalculationEvent:
      type: object
      required:
        - id
        - initiativeId
        - changedBy
        - previousScore
        - newScore
        - previousPriority
        - newPriority
        - changes
        - createdAt
      properties:
        id:
          type: integer
        initiativeId:
          type: integer
        changedBy:
          type: string
        previousScore:
          type: integer
        newScore:
          type: integer
        previousPriority:
          type: string
        newPriority:
          type: string
        changes:
          type: array
          items:
            $ref: "#/components/schemas/ScoringComponentChange"
        createdAt:
          type: string
          format: date-time
...